+++ core/src/main/java/org/geysermc/geyser/util/WebUtils.java
@@ -1,0 +1,69 @@
+/*
+ * Copyright (c) 2026 GeyserMC. http://geysermc.org
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ *
+ * @author GeyserMC
+ * @link https://github.com/GeyserMC/Geyser
+ */
+
+package org.geysermc.geyser.util;
+
+import org.checkerframework.checker.nullness.qual.Nullable;
+import org.geysermc.geyser.GeyserImpl;
+
+import javax.naming.directory.Attribute;
+import javax.naming.directory.InitialDirContext;
+
+public final class WebUtils {
+
+    private WebUtils() {}
+
+    /**
+     * Find a SRV record for the given address
+     *
+     * @param geyser Geyser instance
+     * @param remoteAddress Address to find the SRV record for
+     * @return The SRV record or null if not found
+     */
+    public static String @Nullable [] findSrvRecord(GeyserImpl geyser, String remoteAddress) {
+        try {
+            // Searches for a server address and a port from a SRV record of the specified host name
+            InitialDirContext ctx = new InitialDirContext();
+            Attribute attr = ctx.getAttributes("dns:///_minecraft._tcp." + remoteAddress, new String[]{"SRV"}).get("SRV");
+            // size > 0 = SRV entry found
+            if (attr != null && attr.size() > 0) {
+                return ((String) attr.get(0)).split(" ");
+            }
+        } catch (Exception | NoClassDefFoundError ex) { // Check for a NoClassDefFoundError to prevent Android crashes
+            if (geyser.config().debugMode()) {
+                geyser.getLogger().debug("Exception while trying to find an SRV record for the remote host.");
+                ex.printStackTrace(); // Otherwise we can get a stack trace for any domain that doesn't have an SRV record
+            }
+        }
+        return null;
+    }
+
+    public static String toHttps(String url) {
+        if (url != null && url.startsWith("http://")) {
+            return "https://" + url.substring(7);
+        }
+        return url;
+    }
+}
